# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Teaching Planning Assistant API",
    "description": "Autonomous educational content generation engine for Chilean curriculum",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/v1/validation/validate-content": {
      "post": {
        "tags": [
          "Content Validation"
        ],
        "summary": "Validate Content",
        "description": "Validate educational content against quality criteria.\n\nPerforms comprehensive validation including curriculum alignment,\npedagogical quality, content accuracy, age appropriateness,\nlanguage quality, accessibility, safety, bias checks, and completeness.\n\nReturns validation status (passed/failed/needs_review), detailed\nissues with fix suggestions, and overall quality score.\n\nAuthentication: JWT token OR API Key (with content:validate permission)",
        "operationId": "validate_content_api_v1_validation_validate_content_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidateContentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidateContentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/v1/validation/quick-check": {
      "post": {
        "tags": [
          "Content Validation"
        ],
        "summary": "Quick Check Content",
        "description": "Quick validation check for content (minimal validation level).\n\nA lighter version of validate_content that only checks for\ncritical errors and safety issues. Faster but less thorough.\n\nAuthentication: JWT token OR API Key",
        "operationId": "quick_check_content_api_v1_validation_quick_check_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "content",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Content"
            }
          },
          {
            "name": "content_type",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ContentType"
            }
          },
          {
            "name": "grade_level",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Grade Level"
            }
          },
          {
            "name": "subject",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Subject"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidateContentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/tutor/chat": {
      "post": {
        "tags": [
          "AI Tutor"
        ],
        "summary": "Tutor Chat",
        "description": "Chat with the AI tutor.\n\nProvides personalized educational support through conversation.\nThe tutor can explain concepts, guide practice, ask Socratic\nquestions, provide encouragement, and help with review.\n\nSupports NEE adaptations and different learning styles.\n\nAuthentication: JWT token OR API Key (with tutor:chat permission)",
        "operationId": "tutor_chat_api_v1_tutor_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TutorChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TutorChatResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/v1/tutor/sessions/{session_id}": {
      "get": {
        "tags": [
          "AI Tutor"
        ],
        "summary": "Get Session Info",
        "description": "Get information about a tutoring session.\n\nReturns topics covered and session activity.",
        "operationId": "get_session_info_api_v1_tutor_sessions__session_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "AI Tutor"
        ],
        "summary": "End Session",
        "description": "End a tutoring session.\n\nClears session data.",
        "operationId": "end_session_api_v1_tutor_sessions__session_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChatMessage": {
        "properties": {
          "role": {
            "type": "string",
            "title": "Role"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "role",
          "content"
        ],
        "title": "ChatMessage",
        "description": "A single chat message"
      },
...
      "ContentType": {
        "type": "string",
        "enum": [
          "lesson",
          "quiz",
          "activity",
          "exam",
          "reading",
          "worksheet"
        ],
        "title": "ContentType",
        "description": "Type of content being adapted"
      },
...
      "DifficultyLevel": {
        "type": "string",
        "enum": [
          "easy",
          "medium",
          "hard"
        ],
        "title": "DifficultyLevel"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TutorChatRequest": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Student's message/question"
          },
          "context": {
            "$ref": "#/components/schemas/TutorContext"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id",
            "description": "Existing session ID for continuity"
          },
          "conversation_history": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ChatMessage"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conversation History",
            "description": "Previous messages in conversation (max 10)"
          },
          "tutor_role": {
            "$ref": "#/components/schemas/TutorRole",
            "default": "explain"
          },
          "response_style": {
            "type": "string",
            "title": "Response Style",
            "description": "friendly, formal, playful",
            "default": "friendly"
          },
          "language": {
            "type": "string",
            "title": "Language",
            "default": "es"
          },
          "max_response_length": {
            "type": "integer",
            "title": "Max Response Length",
            "description": "Max words in response",
            "default": 500
          },
          "include_examples": {
            "type": "boolean",
            "title": "Include Examples",
            "default": true
          },
          "include_follow_up_questions": {
            "type": "boolean",
            "title": "Include Follow Up Questions",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "message",
          "context"
        ],
        "title": "TutorChatRequest",
        "description": "Request model for tutor chat"
      },
      "TutorChatResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "response": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TutorResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "conversation_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conversation Summary"
          },
          "topics_covered": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Topics Covered",
            "default": []
          },
          "generation_metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Generation Metadata"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "success",
          "session_id",
          "generation_metadata"
        ],
        "title": "TutorChatResponse",
        "description": "Response model for tutor chat"
      },
      "TutorContext": {
        "properties": {
          "student_grade_level": {
            "type": "string",
            "title": "Student Grade Level"
          },
          "subject": {
            "type": "string",
            "title": "Subject"
          },
          "current_topic": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Topic"
          },
          "linked_oas": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Linked Oas"
          },
          "student_nee": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Student Nee"
          },
          "learning_style": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Learning Style"
          },
          "difficulty_preference": {
            "$ref": "#/components/schemas/DifficultyLevel",
            "default": "medium"
          }
        },
        "type": "object",
        "required": [
          "student_grade_level",
          "subject"
        ],
        "title": "TutorContext",
        "description": "Context for tutoring session"
      },
      "TutorResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          },
          "examples": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Examples"
          },
          "follow_up_questions": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Follow Up Questions"
          },
          "resources": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resources"
          },
          "encouragement": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Encouragement"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "TutorResponse",
        "description": "A single tutor response"
      },
      "TutorRole": {
        "type": "string",
        "enum": [
          "explain",
          "practice",
          "question",
          "encourage",
          "review"
        ],
        "title": "TutorRole",
        "description": "Role/mode for the tutor"
      },
...
      "ValidateContentRequest": {
        "properties": {
          "content_type": {
            "$ref": "#/components/schemas/ContentType",
            "description": "Type of content"
          },
          "content": {
            "type": "string",
            "title": "Content",
            "description": "The content to validate (JSON string or markdown)"
          },
          "content_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Id",
            "description": "ID if from TPA"
          },
          "grade_level": {
            "type": "string",
            "title": "Grade Level",
            "description": "Target grade level"
          },
          "subject": {
            "type": "string",
            "title": "Subject",
            "description": "Subject area"
          },
          "linked_oas": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Linked Oas",
            "description": "OAs the content should align with"
          },
          "validation_level": {
            "type": "string",
            "title": "Validation Level",
            "description": "minimal, standard, strict",
            "default": "standard"
          },
          "categories_to_check": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ValidationCategory"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Categories To Check",
            "description": "Categories to validate (default: all)"
          },
          "minimum_score": {
            "type": "number",
            "title": "Minimum Score",
            "description": "Minimum acceptable overall score",
            "default": 70.0
          },
          "error_threshold": {
            "type": "integer",
            "title": "Error Threshold",
            "description": "Max errors allowed for pass",
            "default": 0
          },
          "warning_threshold": {
            "type": "integer",
            "title": "Warning Threshold",
            "description": "Max warnings allowed for pass",
            "default": 5
          },
          "language": {
            "type": "string",
            "title": "Language",
            "default": "es"
          },
          "include_recommendations": {
            "type": "boolean",
            "title": "Include Recommendations",
            "default": true
          },
          "include_fix_suggestions": {
            "type": "boolean",
            "title": "Include Fix Suggestions",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "content_type",
          "content",
          "grade_level",
          "subject"
        ],
        "title": "ValidateContentRequest",
        "description": "Request model for content validation"
      },
      "ValidateContentResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "Whether validation completed"
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ValidationResult-Output"
              },
              {
                "type": "null"
              }
            ],
            "description": "Validation result"
          },
          "generation_metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Generation Metadata",
            "description": "Metadata"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error",
            "description": "Error if failed"
          }
        },
        "type": "object",
        "required": [
          "success",
          "generation_metadata"
        ],
        "title": "ValidateContentResponse",
        "description": "Response model for content validation"
      },
      "ValidationCategory": {
        "type": "string",
        "enum": [
          "curriculum_alignment",
          "pedagogical_quality",
          "content_accuracy",
          "age_appropriateness",
          "language_quality",
          "accessibility",
          "safety",
          "bias_check",
          "completeness"
        ],
        "title": "ValidationCategory",
        "description": "Categories of validation checks"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "ValidationResult-Output": {
        "properties": {
          "validation_id": {
            "type": "string",
            "title": "Validation Id",
            "description": "Unique validation ID"
          },
          "content_type": {
            "type": "string",
            "title": "Content Type",
            "description": "Type of content validated"
          },
          "status": {
            "$ref": "#/components/schemas/ValidationStatus"
          },
          "overall_score": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Overall Score",
            "description": "Overall quality score"
          },
          "is_publishable": {
            "type": "boolean",
            "title": "Is Publishable",
            "description": "Whether content can be published"
          },
          "checks_performed": {
            "items": {
              "$ref": "#/components/schemas/ValidationCheck-Output"
            },
            "type": "array",
            "title": "Checks Performed",
            "description": "All checks performed"
          },
          "issues_found": {
            "items": {
              "$ref": "#/components/schemas/ValidationIssue"
            },
            "type": "array",
            "title": "Issues Found",
            "description": "All issues consolidated"
          },
          "total_checks": {
            "type": "integer",
            "title": "Total Checks",
            "description": "Total number of checks"
          },
          "passed_checks": {
            "type": "integer",
            "title": "Passed Checks",
            "description": "Number of passed checks"
          },
          "error_count": {
            "type": "integer",
            "title": "Error Count",
            "description": "Number of errors"
          },
          "warning_count": {
            "type": "integer",
            "title": "Warning Count",
            "description": "Number of warnings"
          },
          "category_scores": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Category Scores",
            "description": "Scores by category"
          },
          "recommendations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Recommendations",
            "description": "Prioritized recommendations"
          },
          "estimated_fix_time_minutes": {
            "type": "integer",
            "title": "Estimated Fix Time Minutes",
            "description": "Estimated time to fix all issues"
          },
          "requires_human_review": {
            "type": "boolean",
            "title": "Requires Human Review",
            "default": false
          },
          "review_priority": {
            "type": "string",
            "title": "Review Priority",
            "description": "low, normal, high, critical",
            "default": "normal"
          },
          "review_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Notes"
          }
        },
        "type": "object",
        "required": [
          "validation_id",
          "content_type",
          "status",
          "overall_score",
          "is_publishable",
          "checks_performed",
          "issues_found",
          "total_checks",
          "passed_checks",
          "error_count",
          "warning_count",
          "category_scores",
          "recommendations",
          "estimated_fix_time_minutes"
        ],
        "title": "ValidationResult",
        "description": "Complete validation result"
      },
...
      "ValidationStatus": {
        "type": "string",
        "enum": [
          "passed",
          "passed_with_warnings",
          "failed",
          "needs_review"
        ],
        "title": "ValidationStatus",
        "description": "Overall validation status"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "token"
          }
        }
      },
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  }
}